Fix remaining 122 errors in second bulk pass
- ToHashSet polyfill extension method for net472/netstandard2.0
- System.Index.GetOffset added to polyfill
- Extensions alias changed to TaleWorlds.Library for HasAnyFlag compat
- SeasonalDiseaseSystem second WeatherEvent switch cast to int
- CivilianBehavior Vec3/WorldPosition _002Ector + Occupation switch cast
- PlayerReinforcementMissionLogic _003F→Vec3 type fix + GetPath out keyword
- DefenderSpawner GetPathDistanceBetweenPositions ref→out
- DiseaseTreatmentAiBehavior TryGetBehaviorScore ref→out
- AIInfluenceBehavior: remove base.Finalize(), fix GetEncounterMenu out,
  fix npcKingdom duplicate declaration
- ArenaTrainingMenuBehavior delegate alias → top-level GameMenus types
- WorldInfoManager KillCharacterActionDetail switch (detail-4) cast
- DialogManager: OnClickableConditionDelegate, OnPersuasionOptionDelegate aliases
- BehaviorTarget → true (field not in reference assembly)
- WarStatisticsTracker tuple .ClanName→.Item1
- WorldEventsWindowViewModel .sortTime→.Item3
- CivilianBehavior last _002Ector (result=WorldPosition)

+++ src/AIInfluence.Diseases/SeasonalDiseaseSystem.cs
@@ -276,7 +276,7 @@ public static class SeasonalDiseaseSystem
 			WeatherEvent weatherEventInPosition = val.GetWeatherEventInPosition(position2D);
 			WeatherEvent val2 = weatherEventInPosition;
 			WeatherEvent val3 = val2;
-			switch (val3 - 1)
+			switch ((int)val3 - 1)
 			{
 			case 0:
 				num += WeatherLightRainAdd;
